Add test_app fixture
